ref: move directory from features/publicform/logic to features/logic
