+++ frontend/src/features/logic/utils/typeguards.ts
